# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Store Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  description?: string;
  amount?: number;
  category?: string;
  date?: string;
  notes?: string;
}